+++ lib/layrz_models.freezed.dart
@@ -198,7 +198,7 @@ class _$MonitorRealWaypointImpl implements _MonitorRealWaypoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MonitorRealWaypointImpl &&
@@ -543,7 +543,7 @@ class _$MonitorActiveCheckpointImpl implements _MonitorActiveCheckpoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MonitorActiveCheckpointImpl &&
@@ -837,7 +837,7 @@ class _$LinkShortcutImpl implements _LinkShortcut {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LinkShortcutImpl &&
@@ -1150,7 +1150,7 @@ class _$AccessImpl implements _Access {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AccessImpl &&
@@ -1668,7 +1668,7 @@ class _$ActionImpl implements _Action {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ActionImpl &&
@@ -2038,7 +2038,7 @@ class _$ActionGeofenceSettingsImpl implements _ActionGeofenceSettings {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ActionGeofenceSettingsImpl &&
@@ -2786,7 +2786,7 @@ class _$AlgorithmImpl implements _Algorithm {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AlgorithmImpl &&
@@ -3033,6 +3033,14 @@ mixin _$Asset {
   /// [globalSensors] is the list of global sensors of the asset.
   List<Sensor>? get globalSensors => throw _privateConstructorUsedError;
 
+  /// [mappitLaborStartTime] is the time of the labor start time.
+  @TimeOfDayOrNullConverter()
+  TimeOfDay? get mappitLaborStartTime => throw _privateConstructorUsedError;
+
+  /// [mappitLaborEndTime] is the time of the labor end time.
+  @TimeOfDayOrNullConverter()
+  TimeOfDay? get mappitLaborEndTime => throw _privateConstructorUsedError;
+
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
   $AssetCopyWith<Asset> get copyWith => throw _privateConstructorUsedError;
@@ -3086,7 +3094,9 @@ abstract class $AssetCopyWith<$Res> {
       List<String>? linkedSupplyPointAssetsIds,
       StaticPosition? staticPosition,
       List<String> parameters,
-      List<Sensor>? globalSensors});
+      List<Sensor>? globalSensors,
+      @TimeOfDayOrNullConverter() TimeOfDay? mappitLaborStartTime,
+      @TimeOfDayOrNullConverter() TimeOfDay? mappitLaborEndTime});
 
   $AvatarCopyWith<$Res>? get dynamicIcon;
   $CategoryCopyWith<$Res>? get kind;
@@ -3154,6 +3164,8 @@ class _$AssetCopyWithImpl<$Res, $Val extends Asset>
     Object? staticPosition = freezed,
     Object? parameters = null,
     Object? globalSensors = freezed,
+    Object? mappitLaborStartTime = freezed,
+    Object? mappitLaborEndTime = freezed,
   }) {
     return _then(_value.copyWith(
       id: null == id
@@ -3328,6 +3340,14 @@ class _$AssetCopyWithImpl<$Res, $Val extends Asset>
           ? _value.globalSensors
           : globalSensors // ignore: cast_nullable_to_non_nullable
               as List<Sensor>?,
+      mappitLaborStartTime: freezed == mappitLaborStartTime
+          ? _value.mappitLaborStartTime
+          : mappitLaborStartTime // ignore: cast_nullable_to_non_nullable
+              as TimeOfDay?,
+      mappitLaborEndTime: freezed == mappitLaborEndTime
+          ? _value.mappitLaborEndTime
+          : mappitLaborEndTime // ignore: cast_nullable_to_non_nullable
+              as TimeOfDay?,
     ) as $Val);
   }
 
@@ -3479,7 +3499,9 @@ abstract class _$$AssetImplCopyWith<$Res> implements $AssetCopyWith<$Res> {
       List<String>? linkedSupplyPointAssetsIds,
       StaticPosition? staticPosition,
       List<String> parameters,
-      List<Sensor>? globalSensors});
+      List<Sensor>? globalSensors,
+      @TimeOfDayOrNullConverter() TimeOfDay? mappitLaborStartTime,
+      @TimeOfDayOrNullConverter() TimeOfDay? mappitLaborEndTime});
 
   @override
   $AvatarCopyWith<$Res>? get dynamicIcon;
@@ -3553,6 +3575,8 @@ class __$$AssetImplCopyWithImpl<$Res>
     Object? staticPosition = freezed,
     Object? parameters = null,
     Object? globalSensors = freezed,
+    Object? mappitLaborStartTime = freezed,
+    Object? mappitLaborEndTime = freezed,
   }) {
     return _then(_$AssetImpl(
       id: null == id
@@ -3727,6 +3751,14 @@ class __$$AssetImplCopyWithImpl<$Res>
           ? _value._globalSensors
           : globalSensors // ignore: cast_nullable_to_non_nullable
               as List<Sensor>?,
+      mappitLaborStartTime: freezed == mappitLaborStartTime
+          ? _value.mappitLaborStartTime
+          : mappitLaborStartTime // ignore: cast_nullable_to_non_nullable
+              as TimeOfDay?,
+      mappitLaborEndTime: freezed == mappitLaborEndTime
+          ? _value.mappitLaborEndTime
+          : mappitLaborEndTime // ignore: cast_nullable_to_non_nullable
+              as TimeOfDay?,
     ));
   }
 }
@@ -3777,7 +3809,9 @@ class _$AssetImpl implements _Asset {
       final List<String>? linkedSupplyPointAssetsIds,
       this.staticPosition,
       final List<String> parameters = const [],
-      final List<Sensor>? globalSensors})
+      final List<Sensor>? globalSensors,
+      @TimeOfDayOrNullConverter() this.mappitLaborStartTime,
+      @TimeOfDayOrNullConverter() this.mappitLaborEndTime})
       : _childrenIds = childrenIds,
         _children = children,
         _commands = commands,
@@ -4212,13 +4246,23 @@ class _$AssetImpl implements _Asset {
     return EqualUnmodifiableListView(value);
   }
 
+  /// [mappitLaborStartTime] is the time of the labor start time.
+  @override
+  @TimeOfDayOrNullConverter()
+  final TimeOfDay? mappitLaborStartTime;
+
+  /// [mappitLaborEndTime] is the time of the labor end time.
+  @override
+  @TimeOfDayOrNullConverter()
+  final TimeOfDay? mappitLaborEndTime;
+
   @override
   String toString() {
-    return 'Asset(id: $id, name: $name, plate: $plate, vin: $vin, dynamicIcon: $dynamicIcon, mode: $mode, childrenIds: $childrenIds, children: $children, commands: $commands, referencesIds: $referencesIds, references: $references, authenticatedAssetsIds: $authenticatedAssetsIds, authenticatedAssets: $authenticatedAssets, authenticatedUsersIds: $authenticatedUsersIds, authenticatedUsers: $authenticatedUsers, kindId: $kindId, kind: $kind, customFields: $customFields, sensors: $sensors, primaryId: $primaryId, primary: $primary, devicesIds: $devicesIds, devices: $devices, qrCode: $qrCode, externalIdentifiers: $externalIdentifiers, connection: $connection, telemetry: $telemetry, tags: $tags, access: $access, lastExits: $lastExits, activeTime: $activeTime, contacts: $contacts, mappitLaborHours: $mappitLaborHours, ownerId: $ownerId, loginInfo: $loginInfo, authenticationCard: $authenticationCard, authenticationCardId: $authenticationCardId, nfcIdentifier: $nfcIdentifier, linkedSupplyPointAssets: $linkedSupplyPointAssets, linkedSupplyPointAssetsIds: $linkedSupplyPointAssetsIds, staticPosition: $staticPosition, parameters: $parameters, globalSensors: $globalSensors)';
+    return 'Asset(id: $id, name: $name, plate: $plate, vin: $vin, dynamicIcon: $dynamicIcon, mode: $mode, childrenIds: $childrenIds, children: $children, commands: $commands, referencesIds: $referencesIds, references: $references, authenticatedAssetsIds: $authenticatedAssetsIds, authenticatedAssets: $authenticatedAssets, authenticatedUsersIds: $authenticatedUsersIds, authenticatedUsers: $authenticatedUsers, kindId: $kindId, kind: $kind, customFields: $customFields, sensors: $sensors, primaryId: $primaryId, primary: $primary, devicesIds: $devicesIds, devices: $devices, qrCode: $qrCode, externalIdentifiers: $externalIdentifiers, connection: $connection, telemetry: $telemetry, tags: $tags, access: $access, lastExits: $lastExits, activeTime: $activeTime, contacts: $contacts, mappitLaborHours: $mappitLaborHours, ownerId: $ownerId, loginInfo: $loginInfo, authenticationCard: $authenticationCard, authenticationCardId: $authenticationCardId, nfcIdentifier: $nfcIdentifier, linkedSupplyPointAssets: $linkedSupplyPointAssets, linkedSupplyPointAssetsIds: $linkedSupplyPointAssetsIds, staticPosition: $staticPosition, parameters: $parameters, globalSensors: $globalSensors, mappitLaborStartTime: $mappitLaborStartTime, mappitLaborEndTime: $mappitLaborEndTime)';
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetImpl &&
@@ -4291,7 +4335,11 @@ class _$AssetImpl implements _Asset {
             const DeepCollectionEquality()
                 .equals(other._parameters, _parameters) &&
             const DeepCollectionEquality()
-                .equals(other._globalSensors, _globalSensors));
+                .equals(other._globalSensors, _globalSensors) &&
+            (identical(other.mappitLaborStartTime, mappitLaborStartTime) ||
+                other.mappitLaborStartTime == mappitLaborStartTime) &&
+            (identical(other.mappitLaborEndTime, mappitLaborEndTime) ||
+                other.mappitLaborEndTime == mappitLaborEndTime));
   }
 
   @JsonKey(ignore: true)
@@ -4340,7 +4388,9 @@ class _$AssetImpl implements _Asset {
         const DeepCollectionEquality().hash(_linkedSupplyPointAssetsIds),
         staticPosition,
         const DeepCollectionEquality().hash(_parameters),
-        const DeepCollectionEquality().hash(_globalSensors)
+        const DeepCollectionEquality().hash(_globalSensors),
+        mappitLaborStartTime,
+        mappitLaborEndTime
       ]);
 
   @JsonKey(ignore: true)
@@ -4359,49 +4409,52 @@ class _$AssetImpl implements _Asset {
 
 abstract class _Asset implements Asset {
   const factory _Asset(
-      {required final String id,
-      required final String name,
-      final String? plate,
-      final String? vin,
-      final Avatar? dynamicIcon,
-      @AssetModeOrNullConverter() final AssetMode? mode,
-      final List<String>? childrenIds,
-      final List<Asset>? children,
-      final List<DeviceCommand>? commands,
-      final List<String>? referencesIds,
-      final List<Reference>? references,
-      final List<String>? authenticatedAssetsIds,
-      final List<Asset>? authenticatedAssets,
-      final List<String>? authenticatedUsersIds,
-      final List<User>? authenticatedUsers,
-      final String? kindId,
-      final Category? kind,
-      final List<CustomField>? customFields,
-      final List<Sensor>? sensors,
-      final String? primaryId,
-      final Device? primary,
-      final List<String>? devicesIds,
-      final List<Device>? devices,
-      final String? qrCode,
-      final List<String>? externalIdentifiers,
-      final Connection? connection,
-      final AssetTelemetry? telemetry,
-      final List<Tag>? tags,
-      final List<Access>? access,
-      final List<AtsExit>? lastExits,
-      final String? activeTime,
-      final List<ContactInfo>? contacts,
-      final List<MappitLaborHour>? mappitLaborHours,
-      final String? ownerId,
-      final AssetLoginInfo? loginInfo,
-      final AtsAuthenticationCard? authenticationCard,
-      final String? authenticationCardId,
-      final String? nfcIdentifier,
-      final List<Asset>? linkedSupplyPointAssets,
-      final List<String>? linkedSupplyPointAssetsIds,
-      final StaticPosition? staticPosition,
-      final List<String> parameters,
-      final List<Sensor>? globalSensors}) = _$AssetImpl;
+          {required final String id,
+          required final String name,
+          final String? plate,
+          final String? vin,
+          final Avatar? dynamicIcon,
+          @AssetModeOrNullConverter() final AssetMode? mode,
+          final List<String>? childrenIds,
+          final List<Asset>? children,
+          final List<DeviceCommand>? commands,
+          final List<String>? referencesIds,
+          final List<Reference>? references,
+          final List<String>? authenticatedAssetsIds,
+          final List<Asset>? authenticatedAssets,
+          final List<String>? authenticatedUsersIds,
+          final List<User>? authenticatedUsers,
+          final String? kindId,
+          final Category? kind,
+          final List<CustomField>? customFields,
+          final List<Sensor>? sensors,
+          final String? primaryId,
+          final Device? primary,
+          final List<String>? devicesIds,
+          final List<Device>? devices,
+          final String? qrCode,
+          final List<String>? externalIdentifiers,
+          final Connection? connection,
+          final AssetTelemetry? telemetry,
+          final List<Tag>? tags,
+          final List<Access>? access,
+          final List<AtsExit>? lastExits,
+          final String? activeTime,
+          final List<ContactInfo>? contacts,
+          final List<MappitLaborHour>? mappitLaborHours,
+          final String? ownerId,
+          final AssetLoginInfo? loginInfo,
+          final AtsAuthenticationCard? authenticationCard,
+          final String? authenticationCardId,
+          final String? nfcIdentifier,
+          final List<Asset>? linkedSupplyPointAssets,
+          final List<String>? linkedSupplyPointAssetsIds,
+          final StaticPosition? staticPosition,
+          final List<String> parameters,
+          final List<Sensor>? globalSensors,
+          @TimeOfDayOrNullConverter() final TimeOfDay? mappitLaborStartTime,
+          @TimeOfDayOrNullConverter() final TimeOfDay? mappitLaborEndTime}) =
+      _$AssetImpl;
 
   factory _Asset.fromJson(Map<String, dynamic> json) = _$AssetImpl.fromJson;
 
@@ -4592,6 +4645,16 @@ abstract class _Asset implements Asset {
   /// [globalSensors] is the list of global sensors of the asset.
   List<Sensor>? get globalSensors;
   @override
+
+  /// [mappitLaborStartTime] is the time of the labor start time.
+  @TimeOfDayOrNullConverter()
+  TimeOfDay? get mappitLaborStartTime;
+  @override
+
+  /// [mappitLaborEndTime] is the time of the labor end time.
+  @TimeOfDayOrNullConverter()
+  TimeOfDay? get mappitLaborEndTime;
+  @override
   @JsonKey(ignore: true)
   _$$AssetImplCopyWith<_$AssetImpl> get copyWith =>
       throw _privateConstructorUsedError;
@@ -4723,7 +4786,7 @@ class _$ContactInfoImpl implements _ContactInfo {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ContactInfoImpl &&
@@ -4905,7 +4968,7 @@ class _$AssetLoginInfoImpl implements _AssetLoginInfo {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetLoginInfoImpl &&
@@ -5094,7 +5157,7 @@ class _$StaticPositionImpl implements _StaticPosition {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$StaticPositionImpl &&
@@ -5327,7 +5390,7 @@ class _$AvatarImpl implements _Avatar {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AvatarImpl &&
@@ -6008,7 +6071,7 @@ class _$BillingPlanImpl implements _BillingPlan {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$BillingPlanImpl &&
@@ -6465,7 +6528,7 @@ class _$CareProtocolImpl implements _CareProtocol {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CareProtocolImpl &&
@@ -6769,7 +6832,7 @@ class _$CareTaskImpl implements _CareTask {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CareTaskImpl &&
@@ -6992,7 +7055,7 @@ class _$CommentOwnerImpl implements _CommentOwner {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CommentOwnerImpl &&
@@ -7212,7 +7275,7 @@ class _$CaseCommentImpl implements _CaseComment {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CaseCommentImpl &&
@@ -7627,7 +7690,7 @@ class _$CaseImpl implements _Case {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CaseImpl &&
@@ -7885,7 +7948,7 @@ class _$CategoryImpl implements _Category {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CategoryImpl &&
@@ -8213,7 +8276,7 @@ class _$LayrzChartImpl implements _LayrzChart {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LayrzChartImpl &&
@@ -8561,7 +8624,7 @@ class _$CheckpointImpl implements _Checkpoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CheckpointImpl &&
@@ -8847,7 +8910,7 @@ class _$WaypointImpl implements _Waypoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$WaypointImpl &&
@@ -9037,7 +9100,7 @@ class _$LinkedAssetToCheckpointIdImpl implements _LinkedAssetToCheckpointId {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LinkedAssetToCheckpointIdImpl &&
@@ -9637,7 +9700,7 @@ class _$CloudEntryImpl implements _CloudEntry {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CloudEntryImpl &&
@@ -10289,7 +10352,7 @@ class _$DeviceCommandImpl implements _DeviceCommand {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceCommandImpl &&
@@ -10601,7 +10664,7 @@ class _$ConnectionImpl implements _Connection {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ConnectionImpl &&
@@ -10814,7 +10877,7 @@ class _$CountryImpl implements _Country {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CountryImpl &&
@@ -10991,7 +11054,7 @@ class _$PhoneNumberImpl extends _PhoneNumber {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PhoneNumberImpl &&
@@ -11159,7 +11222,7 @@ class _$NullablePhoneNumberImpl extends _NullablePhoneNumber {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$NullablePhoneNumberImpl &&
@@ -11371,7 +11434,7 @@ class _$CityImpl implements _City {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CityImpl &&
@@ -11570,7 +11633,7 @@ class _$CountryStateImpl implements _CountryState {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CountryStateImpl &&
@@ -11974,7 +12037,7 @@ class _$CredentialFieldImpl implements _CredentialField {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CredentialFieldImpl &&
@@ -12269,7 +12332,7 @@ class _$CustomFieldImpl implements _CustomField {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomFieldImpl &&
@@ -12454,7 +12517,7 @@ class _$CustomFieldInputImpl implements _CustomFieldInput {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomFieldInputImpl &&
@@ -12753,7 +12816,7 @@ class _$DbPartitionImpl implements _DbPartition {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DbPartitionImpl &&
@@ -12990,7 +13053,7 @@ class _$DepartmentImpl implements _Department {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DepartmentImpl &&
@@ -13564,7 +13627,7 @@ class _$DeviceImpl implements _Device {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceImpl &&
@@ -13916,7 +13979,7 @@ class _$EmailTemplateImpl implements _EmailTemplate {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$EmailTemplateImpl &&
@@ -14396,7 +14459,7 @@ class _$EmployeeImpl implements _Employee {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$EmployeeImpl &&
@@ -14676,7 +14739,7 @@ class _$FtpAccountImpl implements _FtpAccount {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$FtpAccountImpl &&
@@ -15227,7 +15290,7 @@ class _$LayrzFunctionImpl implements _LayrzFunction {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LayrzFunctionImpl &&
@@ -15586,7 +15649,7 @@ class _$GenericPermissionItemImpl implements _GenericPermissionItem {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GenericPermissionItemImpl &&
@@ -16110,7 +16173,7 @@ class _$GenericPermissionImpl implements _GenericPermission {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GenericPermissionImpl &&
@@ -16325,7 +16388,7 @@ class _$GeofencePointImpl implements _GeofencePoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeofencePointImpl &&
@@ -16861,7 +16924,7 @@ class _$GeofenceImpl implements _Geofence {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeofenceImpl &&
@@ -17464,7 +17527,7 @@ class _$GeoPointImpl implements _GeoPoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeoPointImpl &&
@@ -17672,7 +17735,7 @@ class _$GuideImpl implements _Guide {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GuideImpl &&
@@ -17856,7 +17919,7 @@ class _$LanguageImpl implements _Language {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LanguageImpl &&
@@ -18086,7 +18149,7 @@ class _$LintErrorImpl implements _LintError {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LintErrorImpl &&
@@ -18376,7 +18439,7 @@ class _$ModelImpl implements _Model {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ModelImpl &&
@@ -19055,7 +19118,7 @@ class _$OperationImpl implements _Operation {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$OperationImpl &&
@@ -19378,7 +19441,7 @@ class _$HttpHeaderImpl implements _HttpHeader {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$HttpHeaderImpl &&
@@ -20316,7 +20379,7 @@ class _$PresetImpl implements _Preset {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PresetImpl &&
@@ -20588,7 +20651,7 @@ class _$ProfileImpl implements _Profile {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ProfileImpl &&
@@ -20872,7 +20935,7 @@ class _$ReferenceImpl implements _Reference {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReferenceImpl &&
@@ -21249,7 +21312,7 @@ class _$ReportTemplateImpl implements _ReportTemplate {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplateImpl &&
@@ -21541,7 +21604,7 @@ class _$ReportTemplatePageImpl implements _ReportTemplatePage {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplatePageImpl &&
@@ -21768,7 +21831,7 @@ class _$ReportTemplateColImpl implements _ReportTemplateCol {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplateColImpl &&
@@ -21976,7 +22039,7 @@ class _$CustomReportImpl implements _CustomReport {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomReportImpl &&
@@ -22903,7 +22966,7 @@ class _$SensorImpl implements _Sensor {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SensorImpl &&
@@ -23373,7 +23436,7 @@ class _$MaskPointImpl implements _MaskPoint {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MaskPointImpl &&
@@ -23552,7 +23615,7 @@ class _$SensorPairImpl implements _SensorPair {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SensorPairImpl &&
@@ -23805,7 +23868,7 @@ class _$SuspendedServiceImpl implements _SuspendedService {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SuspendedServiceImpl &&
@@ -25433,7 +25496,7 @@ class _$TagImpl implements _Tag {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TagImpl &&
@@ -26072,7 +26135,7 @@ class _$DeviceTelemetryImpl extends _DeviceTelemetry {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceTelemetryImpl &&
@@ -26417,7 +26480,7 @@ class _$AssetTelemetryImpl extends _AssetTelemetry {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetTelemetryImpl &&
@@ -26754,7 +26817,7 @@ class _$TelemetryPositionImpl extends _TelemetryPosition {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TelemetryPositionImpl &&
@@ -26966,7 +27029,7 @@ class _$TelemetrySensorImpl extends _TelemetrySensor {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TelemetrySensorImpl &&
@@ -27192,7 +27255,7 @@ class _$TimezoneImpl implements _Timezone {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TimezoneImpl &&
@@ -27364,7 +27427,7 @@ class _$TokenImpl implements _Token {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TokenImpl &&
@@ -28257,7 +28320,7 @@ class _$TriggerImpl implements _Trigger {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TriggerImpl &&
@@ -28780,7 +28843,7 @@ class _$TriggerActivationImpl implements _TriggerActivation {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TriggerActivationImpl &&
@@ -29895,7 +29958,7 @@ class _$TripImpl extends _Trip {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TripImpl &&
@@ -30153,7 +30216,7 @@ class _$PointImpl extends _Point {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PointImpl &&
@@ -30343,7 +30406,7 @@ class _$ConciergeFormImpl implements _ConciergeForm {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ConciergeFormImpl &&
@@ -31250,7 +31313,7 @@ class _$UserImpl implements _User {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$UserImpl &&