+++ lib/src/brickhouse/brickhouse.freezed.dart
@@ -660,7 +660,7 @@ class _$BrickhouseAlertImpl implements _BrickhouseAlert {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$BrickhouseAlertImpl &&
@@ -2017,7 +2017,7 @@ class _$BrickHouseWorkspaceInputImpl implements _BrickHouseWorkspaceInput {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$BrickHouseWorkspaceInputImpl &&