+++ lib/src/converters/converters.dart
@@ -1,6 +1,6 @@
 library converters;
 
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_models/utils.dart';
 import 'package:material_design_icons_flutter/icon_map.dart';
@@ -11,3 +11,4 @@ part 'src/color.dart';
 part 'src/duration.dart';
 part 'src/icon.dart';
 part 'src/timestamp.dart';
+part 'src/time_of_day.dart';